requirements.txt
websockets>=12.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytz>=2024.1
python-socks[asyncio]>=2.4.0
aiohttp_socks>=0.8.0
python-telegram-bot>=20.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
# Step5 micro CHoCH 检测(joshyattridge)
smartmoneyconcepts>=0.0.27
pandas>=2.0.0
numpy>=1.24.0
